Rewrite Mod in Kotlin (#149)
This pull request has unbelievably breaking changes. It rewrites almost all of the files, with the exception of Mixins and UriScheme, to use the Kotlin Language instead of Java. This decision was made since Kotlin handles nulls very well, which was an issue we had in Java.

This rewrite also includes a change in a lot of internal systems, including a rewrite of the module loading system, which is now starting to get support for python.