Frames+ compatibility
For whatever reason, when ct.js runs with Frames+ installed, the
particle injection crashes with a class verification error. Due to the
already uncommon use of the spawnParticle trigger, we just remove the
injection in environments with F+ (and print a warning to console if the
user attempts to register a spawnParticle trigger).