Fix addPage, setPage, and display Methods
The addPage and setPage methods previously would crash due to a cast exception. This commit adds to the NBTTagCompound wrapper by adding support for tag lists. I'm really not sure I did this in the best way for the get method because it feels silly to add a whole argument for one specific case. Also the display function seemed to not properly update so that was fixed.