Fix packetSent trigger causing Failed Assertion Errors
If the packet in question is not in the correct thread a scheduled task is created to make it execute in the correct one. Trying to access the packet before this has happened seems to cause Failed Assertion Errors. This commit also changes packetReceived to be in the same place as packetSent. Unfortunately, the amount of injections is still two because some packets are sent already in the correct thread.