Fix triggers not registering properly
Triggers were recently updated to be stored in a concurrent set, but
during this update, a custom comparator was created for triggers which
didn't properly handle two triggers with the same priority, leading
to only one trigger per priority being inserted. This behavior was
already properly handled in OnTrigger#compareTo so now we default
back to that function.