All: Get rid of most of the Kotlin typealiases
Most of these can be handled by the preprocessor. Some of them were
only necessary because they had the same name as the CT class. In
these cases, the typealias was moved to that class. A couple
typealiases have been kept because they are used throughout the
project, and removing them would cause confusion (i.e. "is this
type the CT type or the MC type?").