Misc: Fix registry methods, clean up Item and Block toStrings
BlockType::getRegistryName was returning the id instead of the name
Item::getRegistryName was just erroring
World::getBiome now returns the biome correctly
Verified & simplified PotionEffect::isDurationMax
Server::getPing now only returns 5 when in singleplayer. This will now
return -1 when on title screen, when it used to return 5 there as well.