Renderer: Convert drawStringInWorld to 1.17
Also moved RenderWorld trigger to where it injects for Fabric.
Changed the scaling to be 20x smaller. Before, a scale of 1 was enormous, and now it is very similar to a regular nametag's size.
Also fixes the issue of having 2 vastly different text sizes depending on whether increase is toggled or not.