JSLoader: Do not use file paths in module IDs
In CTRequire::loadCTModule, we were taking the name of the module and
the name of the entry point in that module (typically just "index.js"),
then combining these and passing it off to Rhino to load. However, Rhino
located the module on the filesystem with the URI, not the name, so
including the entry point in the module ID is unnecessary.

The module ID is what Rhino uses to cache modules. Combining the name
and entry point is fine as long as this matches the behavior of the rest
of the engine. When Rhino processes an import, the ID is of course just
the string given to the import, so there is no entry point. This caused
modules imported based on the "requires" key of a metadata.json file and
modules imported with "import" to have different IDs and thus be
evaluated multiple times.

This commit simply removes the entry parameter and just uses the name as
the ID, which matches the normal module loading behavior present in
Rhino.

Fixes #379