JSLoader: Do not try to hold multiple contexts in a thread
Rhino's context system is really only a wrapper around some internal
thread-local engine state. It's not super important for us to manage in
such a fine degree. What's much more important are the scope objects,
which are shared by default.

The main issue with our previous approach is that when the user executes
a "/ct load", we reload the mod from a background thread. This associates
the "moduleContext" with that background thread, making future uses of it
invalid. The fix is to get rid of all the stored contexts and just call
"enterContext" everywhere. The contexts are caches in a thread-local, so
it isn't that expensive. We can also batch certain calls in the future.